requirements.txt
Django
djangorestframework
Pillow
python-decouple
psycopg2-binary